Move SearchExperiments endpoint (#68)
