+++ .golangci.yml
@@ -9,6 +9,7 @@ linters:
     - gosimple
     - lll
   disable:
+    - ireturn 
     - depguard
     - gochecknoglobals # Immutable globals are fine.
     - exhaustruct # Often the case for protobuf generated code or gorm structs.
