+++ mlflow_go/store/tracking.py
@@ -25,12 +25,15 @@ from mlflow.protos.service_pb2 import (
     LogParam,
     RestoreExperiment,
     RestoreRun,
+    SearchExperiments,
     SearchRuns,
     SetTag,
     SetTraceTag,
     UpdateExperiment,
     UpdateRun,
 )
+from mlflow.store.entities import PagedList
+from mlflow.store.tracking import SEARCH_MAX_RESULTS_DEFAULT
 from mlflow.utils.uri import resolve_uri_if_local
 
 from mlflow_go import is_go_enabled
@@ -200,6 +203,27 @@ class _TrackingStore:
             key=key,
         )
         self.service.call_endpoint(get_lib().TrackingServiceDeleteTraceTag, request)
+
+    def search_experiments(
+        self,
+        view_type=ViewType.ACTIVE_ONLY,
+        max_results=SEARCH_MAX_RESULTS_DEFAULT,
+        filter_string=None,
+        order_by=None,
+        page_token=None,
+    ):
+        request = SearchExperiments(
+            view_type=view_type,
+            max_results=max_results,
+            filter=filter_string,
+            order_by=order_by,
+            page_token=page_token,
+        )
+        response = self.service.call_endpoint(get_lib().TrackingServiceSearchExperiments, request)
+        experiments = [
+            Experiment.from_proto(proto_experiment) for proto_experiment in response.experiments
+        ]
+        return PagedList(experiments, (response.next_page_token or None))
 
     def set_tag(self, run_id, tag):
         request = SetTag(run_id=run_id, key=tag.key, value=tag.value)
