+++ pkg/tracking/store/sql/runs_internal_test.go
@@ -1,4 +1,3 @@
-//nolint:ireturn
 package sql
 
 import (
